update readme for zoom control, refactor components, remove default export
