+++ src/controls/ZoomControl.tsx
